Implement knn adjacency list graph
To implement a minimum spanning tree algorithm,
we needed to create an adjacency list data
structure suitable for point clouds. Point clouds
typically have no connectivity information, so
we use the k nearest neighbours as topology for
point clouds in many algorithms. To do normals
estimation, we needed the minimum spanning tree
algorithm to run on the knn topology. To prevent
having to copy points and wasting memory, we
added support for the PointView concept, which
is a non-owning type for Points. Our point_view_t
only needs memory space for a single pointer.
The octree, aabb, sphere and geometry functions
needed to be updated to support PointView
instead of Point when ownership was not needed.
We were able to relax type requirements in many
places and octrees of PointView are now possible.
Thus, you could provide any type satisfying
PointView and the octree should be able to store
it. One such use will be to store additional
data at octree nodes, such as tangent planes
and normals, density, etc.