Implement a vector-based dynamic adjacency list
The minimal spanning tree algorithm needs to build the tree dynamically and return it. To do so, we needed a mutable directed graph, where we could add vertices and edges dynamically. We provide a MutableDirectedGraph concept so that the minimal spanning tree algorithm can receive any graph satisfying the concept. We provide a default implementation of a mutable adjacency list that stores vertices in a std::vector and stores connectivity information based on indices to the vector elements.