Build undirected graph from k nearest neighbors
A knn graph type was judged unnecessary. Instead, we expose a function which takes in a KnnSearcher and some vertices and builds an adjacency list from it with bidirectional edges. We create a basic_vertex_t type acting as a view over a point and also adds an identifier to it. Any vertex to be used in graph algorithms should now satisfy the GraphVertex concept so that we can identify vertices by their id. So that we can use inheritance and composition between Point/PointView types and GraphVertex types, we remove many operator overloads from our point types and instead refer to point_predicates.hpp for functions used for point comparison. All algorithms that need point predicates should use the common/ folder. Now that we have a way to construct undirected knn graphs, we can now proceed to implement normals estimation in the library.