Estimate normal with range of points
To implement tangent plane estimation, we need a way to estimate normals using points only. One way to do that is to use PCA (Principal Component Analysis), where we perform eigen decomposition of the covariance matrix of the points, and choose the eigenvector of smallest eigenvalue as the estimated normal. We also take the chance to refine our Point and Normal concepts to require negating them.