Implement depth-first and breadth-first graph searches
We remove unnecessary move ctor and move assign constraints on the PointView concept. move semantics should be an opt-in. We added a test for dfs and bfs and noticed that getting the out edges of a vertex could fail if the vertex actually doesn't have any edges, so we fixed that.