Implement normal estimation of a point cloud
We expose an algorithm/ folder for normal estimation which computes normals and also orients the normals consistently. These algorithms are estimations and not exact solutions as there are no exact solutions for normal estimation in general. We expose the algorithms in a similar style as STL algorithms to support various types of sequences as input/output to the algorithms.