Implement surface nets isosurface extraction
To implement our surface reconstruction algorithm, we use
tangent plane estimation and normal orientation propagation
to define a signed distance function from arbitrary points
in 3d as the distance of a point to the nearest tangent plane.

To do so, we introduce common structures like triangles and
regular 3d grids as discretization objects for our domain and
our SDF from which the isosurface will be our surface.

We relax the requirement that nearest neighbour searches on
our octree can only take in a point of the same type as
the type of the elements the octree stores. In reality, to
perform nearest neighbour searches, we only need a point,
no matter what its type is. The nearest neighbour search
on our octree can now take an arbitrary PointView.

Our lack of use of vector3d was also a problem. A difference
between points p1 and p2 (p2 - p1) should return a vector3d.
Additions should be done between points and vectors too, as
p1 + v1. So we enforce this in the type system.

To output meshes, we also added a write_ply overload taking
vertices and triangles to output ply meshes.

A basic timer class is created to monitor operation execution
times.