Use libigl/imgui for visualization in examples
Also noteworthy that there was a bug in ply io for writing meshes where we forgot to include the uchar (number of indices) for each face. The ply mesh writing is not tested, so it's definitely a todo to test it.

In this commit, we don't yet add the algorithms for computing point cloud densities, filtering, noise, etc, but they are implemented in the examples. We'll add them to the library in future commits in the form of common algorithms.