Define common property maps for pcp
pcp is undergoing refactoring changes to its interface to relax type restrictions in its generic interface. Instead of restricting types to have a certain interface, we now provide property maps to decouple input types from algorithms and data structures. In the logical next step of this commit, algorithms and data structures should now take in any type of element, and ask for property maps whenever it needs to access specific properties of the element.