Use doxygen grouping to automate API reference doc
We use doxygen's grouping feature to define modules
and include classes/structs, functions, typedefs,
statics, etc. in their corresponding module.

Using Breathe directives, we display the auto
generated module documentation in readthedocs
style. Many parts of the documentation are not
filled in yet, so we'll definitely have to do
that in the next commit, at least partially.

The module grouping is a good start, but is
still hard to read, since it includes too
many things. We should use nested groupings
to simplify readability. For example,
in the i/o module, we can define ply and obj
groups to provide further structure to the
documentation, enabling better readability.