Document pcp's introduction and user guide
We wanted to write beautiful documentation
that help users understand our library
better. The overview and feature list were
written with that goal in mind. We also
migrated our markdown README.md documentation
to reStructuredText to avoid duplication.
We didn't want to have to maintain two copies
that could end up unsynchronized. There is
still the examples section of the User Guide
and the design section of the API that need
to be filled in.