Refactor octree API to support property maps
We enable our linked octree data structure to now support storing any element type. To do so, we simply
require users to give us property maps whenever it's relevant. The tests, documentation, examples and benchmarks are updated accordingly, because this is an API breaking change. Graphs and algorithms will need to be refactored to support property maps too in following commits.