Refactor adjacency lists to use index maps
In the process of this refactoring, we realized that we could actually implement the adjacency lists as multimaps. Since we now use index maps, it's trivial to feed them to std::unordered_multimap as its hasher and key equality comparator. We adjust the tests to use the new interface requiring index maps. The prim MST implementation has been commented out for now. It isn't used in our codebase and we haven't adjusted it to use index maps. We might have to make that change later down the line, especially if new features need MSTs.