Update normals estimation to use property maps
The need for a vertex type, i.e. a point with index, is pretty
much deprecated. We modify the normals estimation example
executable to demonstrate the use of property maps instead
of vertex types. In doing so, we have had to update many
interfaces in the common module for property map support,
like geometric center computations, normals estimations, and
tangent plane estimations. As such, we update parts of the code that
depend on these common functions.