Fix adjacency list edge traversal/access test case
The edge traversal/access test case depended on the order
in which edges were stored for a given hash map key.
MSVC's unordered_multimap implementation stored the edges in
ascending order whereas clang's implementation stored the edges
in descending order for the same key. We thus changed the test
case to disregard order of traversal.

At the same time, we take the time to fix compilation warnings
for unix (gcc/clang) builds. We mark the range-v3 library headers
as system headers to silence their warnings.