Specify const qualifiers
When taking a reference to const kd-tree or pointer to const kd-tree or simply a const kd-tree, we wouldn't be able to call the range_search method, because it would violate constness. Since range_search does not modify internal state, we should/have to mark it const, or we won't be able to use it anywhere a kd-tree is marked as const.