Implement bilateral filter for points
We use a bilteral filter with gaussians as our spatial weight function and influence weight function and use a kd-tree for optimized range searches in the support region of the spatial weight function.

Our bilateral filter formulation is the same as in
'Jones, Thouis R., Fredo Durand, and Matthias Zwicker. "Normal improvement for point rendering." IEEE Computer Graphics and Applications 24.4 (2004): 53-56.'