Implement resampling away from edges for EAR
The normal smoothing preprocessing step works well on the fandisk example. The modified anisotropic LOP operator defined with normal awareness as proposed in the edge aware resampling paper does not produce expected results, however. We should look into this problem. Is this a parameter tuning problem, an implementation problem, or other?